pharos-ai-fastapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.21
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
feedparser==6.0.11
trafilatura==1.12.2
newspaper3k==0.2.8
goose3==3.1.19
readability-lxml==0.8.1
boilerpy3==1.0.7
python-dateutil==2.9.0
celery==5.4.0
redis==5.1.1
flower==2.0.1
openai==1.54.3
langgraph>=0.2.50
langchain-openai>=0.2.10
langchain-core>=0.3.21,<0.4.0
requests==2.32.3
urllib3==2.2.3
lxml==5.3.0
Pillow==11.0.0
